backend/pyproject.toml
[project]
name = "dingkwang-site-backend"
version = "0.1.0"
description = "Backend API for Dingkang Wang's personal homepage chatbot"
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "fastapi",
    "uvicorn[standard]",
    "sse-starlette",
    "python-dotenv",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
